Fix stale jobs

+++ controllers/controllers_test.go
@@ -12,7 +12,7 @@ func TestDate(t *testing.T) {
 		startDate := time.Now()
 		endDate := startDate.Add(time.Minute)
 
-		println("Diff seconds", endDate.Sub(startDate).Minutes())
+		println("Diff seconds", int64(endDate.Sub(startDate).Minutes()) )
 	}
 
 }
